[10021] Move item disenchanting static req. checks to server start.

+++ src/game/ItemHandler.cpp
@@ -414,13 +414,13 @@ void WorldSession::HandleItemQuerySingleOpcode( WorldPacket & recv_data )
             data << pProto->Socket[s].Color;
             data << pProto->Socket[s].Content;
         }
-        data << pProto->socketBonus;
-        data << pProto->GemProperties;
-        data << pProto->RequiredDisenchantSkill;
-        data << pProto->ArmorDamageModifier;
-        data << pProto->Duration;                           // added in 2.4.2.8209, duration (seconds)
-        data << pProto->ItemLimitCategory;                  // WotLK, ItemLimitCategory
-        data << pProto->HolidayId;                          // Holiday.dbc?
+        data << uint32(pProto->socketBonus);
+        data << uint32(pProto->GemProperties);
+        data << int32(pProto->RequiredDisenchantSkill);
+        data << float(pProto->ArmorDamageModifier);
+        data << uint32(pProto->Duration);                   // added in 2.4.2.8209, duration (seconds)
+        data << uint32(pProto->ItemLimitCategory);          // WotLK, ItemLimitCategory
+        data << uint32(pProto->HolidayId);                  // Holiday.dbc?
         SendPacket( &data );
     }
     else
